order_mgmt/requirements.txt
Flask
gunicorn
pymysql
Flask-Session
python-dotenv
cryptography
